requirements.txt
Django
dj-database-url
django-environ
psycopg2
whitenoise
gunicorn
django-bootstrap3
djangorestframework
requests
drf-spectacular